Police (#8)
* Added alloggiati service

* Removed entity from tile

* Integrating service into DI

* Implemented session service for AlloggiatiService

* Added Ricevuta get to AlloggiatiService session

* Implemented service table readers

* Implemented TrackedRecord model

* Implemented tracked record serializer

* Implemented download ricevuta method

* Added police nations to db context

* Separated assigned rooms and stays composition into two services

* Implemented assigned bookings recomposition by only assigned

* Iperbooking room entity is now generic

* Assigned room can be generic

* Assigned booking is generic

* Reservation id is long

* Implemented method for excluding empty guest data

* Added data to iperbooking guest model

* Sanitizing name and surname on set

* Removed BookingWithGuestsExtensions

* Static GetSanitizedName

* Document type in tracked record is enum

* Implemented nation converter

* Implemented places converter

* Implemented accomodated type solver

* Implemented composition of tracked record strings

* Implemented test method

* Extracted composing records stuff into a method

* Implemented send method

* Extracted tracked record composing logic into a service

* Preloading room assignments to reduce db calls amount

* Extracted booking with guests composition into IBookingWithGuestsProvider service

* Extracted nation converter creation into a INationConverterProvider service

* Extracted place converter creation into a IPlaceConverterProvider service

* Executing in parallel some tasks

* Tracked records composer composes strings instead of objects

* Removed unused things

+++ BookingCalendarApi/Models/Iperbooking/Guests/Reservation.cs
@@ -2,12 +2,12 @@ namespace BookingCalendarApi.Models.Iperbooking.Guests
 {
     public class Reservation
     {
-        public Reservation(int reservationId)
+        public Reservation(long reservationId)
         {
             ReservationId = reservationId;
         }
 
-        public int ReservationId { get; set; }
+        public long ReservationId { get; set; }
         public IEnumerable<Guest> Guests { get; set; } = new List<Guest>();
     }
 }
